move exception.py
