+++ serverside_challenge_1/challenges/yukikawamura/electricity-rate-simulator/electricity_rate_simulator/exception.py
