+++ serverside_challenge_1/challenges/yukikawamura/electricity-rate-simulator/tests/test_core.py
@@ -1,6 +1,6 @@
 import pytest
 from electricity_rate_simulator.core.electric_simulate import ElectricSimulator
-from electricity_rate_simulator.core.exception import (
+from electricity_rate_simulator.exception import (
     ElectricSimulateClientError,
     ElectricSimulateProviderError,
     ElectricSimulationError,
